LayoutTests/storage/indexeddb/modern/deletedatabase-null-name-exception.html

This test calls deleteDatabase on window.indexedDB with a null database name, making sure there is an exception.
<script>

if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
}

function done()
{
    alert("Done");
    if (window.testRunner)
        testRunner.notifyDone();
}

try {
	var request = window.indexedDB.deleteDatabase();
} catch (e) {
	alert(e);
}

done();
</script>
