LayoutTests/storage/indexeddb/modern/deletedatabase-null-name-exception-expected.txt

ALERT: TypeError: Not enough arguments
ALERT: Done
This test calls deleteDatabase on window.indexedDB with a null database name, making sure there is an exception.
